+++ src/Sulu/Bundle/ContentBundle/DependencyInjection/SuluContentExtension.php
@@ -148,6 +148,7 @@ class SuluContentExtension extends Extension implements PrependExtensionInterfac
         $loader->load('document.xml');
         $loader->load('serializer.xml');
         $loader->load('export.xml');
+        $loader->load('import.xml');
         $loader->load('command.xml');
         $loader->load('link-tag.xml');