+++ src/Sulu/Bundle/ContentBundle/Tests/Unit/Content/Structure/ExcerptStructureExtensionTest.php
@@ -17,8 +17,8 @@ use Sulu\Bundle\SearchBundle\Search\Factory;
 use Sulu\Component\Content\Compat\StructureInterface;
 use Sulu\Component\Content\Compat\StructureManagerInterface;
 use Sulu\Component\Content\ContentTypeManagerInterface;
-use Sulu\Component\Content\Export\ContentExportManagerInterface;
-use Sulu\Component\Content\Import\ContentImportManagerInterface;
+use Sulu\Component\Export\Manager\ExportManagerInterface;
+use Sulu\Component\Import\Manager\ImportManagerInterface;
 
 class ExcerptStructureExtensionTest extends \PHPUnit_Framework_TestCase
 {
@@ -38,14 +38,14 @@ class ExcerptStructureExtensionTest extends \PHPUnit_Framework_TestCase
         $contentTypeManager = $this->prophesize(ContentTypeManagerInterface::class);
         $factory = $this->prophesize(Factory::class);
         $node = $this->prophesize(NodeInterface::class);
-        $contentExportManager = $this->prophesize(ContentExportManagerInterface::class);
-        $contentImportManager = $this->prophesize(ContentImportManagerInterface::class);
+        $exportManager = $this->prophesize(ExportManagerInterface::class);
+        $importManager = $this->prophesize(ImportManagerInterface::class);
 
         $excerptExtension = new ExcerptStructureExtension(
             $structureManager->reveal(),
             $contentTypeManager->reveal(),
-            $contentExportManager->reveal(),
-            $contentImportManager->reveal(),
+            $exportManager->reveal(),
+            $importManager->reveal(),
             $factory->reveal()
         );
 
@@ -68,14 +68,14 @@ class ExcerptStructureExtensionTest extends \PHPUnit_Framework_TestCase
         $contentTypeManager = $this->prophesize(ContentTypeManagerInterface::class);
         $factory = $this->prophesize(Factory::class);
         $node = $this->prophesize(NodeInterface::class);
-        $contentExportManager = $this->prophesize(ContentExportManagerInterface::class);
-        $contentImportManager = $this->prophesize(ContentImportManagerInterface::class);
+        $exportManager = $this->prophesize(ExportManagerInterface::class);
+        $importManager = $this->prophesize(ImportManagerInterface::class);
 
         $excerptExtension = new ExcerptStructureExtension(
             $structureManager->reveal(),
             $contentTypeManager->reveal(),
-            $contentExportManager->reveal(),
-            $contentImportManager->reveal(),
+            $exportManager->reveal(),
+            $importManager->reveal(),
             $factory->reveal()
         );