+++ src/Sulu/Bundle/SnippetBundle/Command/SnippetImportCommand.php
@@ -9,9 +9,9 @@
  * with this source code in the file LICENSE.
  */
 
-namespace Sulu\Bundle\ContentBundle\Command;
+namespace Sulu\Bundle\SnippetBundle\Command;
 
-use Sulu\Component\Content\Import\WebspaceInterface;
+use Sulu\Component\Snippet\Import\SnippetImport;
 use Symfony\Bundle\FrameworkBundle\Command\ContainerAwareCommand;
 use Symfony\Component\Console\Input\InputArgument;
 use Symfony\Component\Console\Input\InputInterface;
@@ -20,47 +20,37 @@ use Symfony\Component\Console\Output\OutputInterface;
 use Symfony\Component\Console\Question\ConfirmationQuestion;
 
 /**
- * Export a webspace in a specific format.
+ * Import snippets translations in a specific format.
  */
-class WebspaceImportCommand extends ContainerAwareCommand
+class SnippetImportCommand extends ContainerAwareCommand
 {
     protected function configure()
     {
-        $this->setName('sulu:webspaces:import')
+        $this->setName('sulu:snippet:import')
             ->addArgument('file', InputArgument::REQUIRED, 'test.xliff')
-            ->addArgument('webspace', InputArgument::REQUIRED)
             ->addArgument('locale', InputArgument::REQUIRED)
             ->addOption('format', 'f', InputOption::VALUE_REQUIRED, '', '1.2.xliff')
-            ->addOption('uuid', 'u', InputOption::VALUE_REQUIRED)
-            ->addOption('exportSuluVersion', '', InputOption::VALUE_OPTIONAL, '1.2 or 1.3', '1.3')
-            ->addOption('overrideSettings', 'o', InputOption::VALUE_OPTIONAL, 'Override Settings-Tab', 'false')
-            ->setDescription('Import webspace');
+            ->setDescription('Import Snippets');
     }
 
     protected function execute(InputInterface $input, OutputInterface $output)
     {
-        $webspaceKey = $input->getArgument('webspace');
         $filePath = $input->getArgument('file');
+
         if (!strpos($filePath, '/') === 0) {
             $filePath = getcwd() . '/' . $filePath;
         }
+
         $locale = $input->getArgument('locale');
         $format = $input->getOption('format');
-        $uuid = $input->getOption('uuid');
-        $overrideSettings = $input->getOption('overrideSettings');
-        $exportSuluVersion = $input->getOption('exportSuluVersion');
 
         $output->writeln([
-            '<info>Language Import</info>',
+            '<info>Language Snippet Import</info>',
             '<info>===============</info>',
             '',
             '<info>Options</info>',
-            'Webspace: ' . $webspaceKey,
             'Locale: ' . $locale,
             'Format: ' . $format,
-            'UUID: ' . $uuid,
-            'Override Setting: ' . $overrideSettings,
-            'Sulu Version by export: ' . $exportSuluVersion . '.x',
             '---------------',
             '',
         ]);
@@ -76,18 +66,14 @@ class WebspaceImportCommand extends ContainerAwareCommand
 
         $output->writeln('<info>Continue!</info>');
 
-        /** @var WebspaceInterface $webspaceImporter */
-        $webspaceImporter = $this->getContainer()->get('sulu_content.import.webspace');
+        /** @var SnippetImport $webspaceImporter */
+        $webspaceImporter = $this->getContainer()->get('sulu_snippet.import.snippet');
 
         $import = $webspaceImporter->import(
-            $webspaceKey,
             $locale,
             $filePath,
             $output,
-            $format,
-            $uuid,
-            $overrideSettings,
-            $exportSuluVersion
+            $format
         );
 
         if ($output->getVerbosity() >= OutputInterface::VERBOSITY_VERBOSE) {