+++ src/Sulu/Bundle/SnippetBundle/DependencyInjection/SuluSnippetExtension.php
@@ -102,6 +102,8 @@ class SuluSnippetExtension extends Extension implements PrependExtensionInterfac
         $loader = new XmlFileLoader($container, new FileLocator(__DIR__ . '/../Resources/config'));
         $loader->load('content.xml');
         $loader->load('snippet.xml');
+        $loader->load('export.xml');
+        $loader->load('import.xml');
         $loader->load('admin.xml');
     }
 }