+++ src/Sulu/Component/Content/Export/WebspaceExport.php
@@ -15,12 +15,11 @@ use Sulu\Bundle\ContentBundle\Document\BasePageDocument;
 use Sulu\Bundle\ContentBundle\Document\PageDocument;
 use Sulu\Bundle\DocumentManagerBundle\Bridge\DocumentInspector;
 use Sulu\Component\Content\Compat\StructureManagerInterface;
-use Sulu\Component\Content\Document\Structure\PropertyValue;
 use Sulu\Component\Content\Extension\ExportExtensionInterface;
 use Sulu\Component\Content\Extension\ExtensionManagerInterface;
-use Sulu\Component\Content\Metadata\BlockMetadata;
-use Sulu\Component\Content\Metadata\PropertyMetadata;
 use Sulu\Component\DocumentManager\DocumentManager;
+use Sulu\Component\Export\Export;
+use Sulu\Component\Export\Manager\ExportManagerInterface;
 use Symfony\Component\Console\Helper\ProgressBar;
 use Symfony\Component\Console\Output\NullOutput;
 use Symfony\Component\Templating\EngineInterface;
@@ -28,18 +27,8 @@ use Symfony\Component\Templating\EngineInterface;
 /**
  * Export Content by given locale to xliff file.
  */
-class Webspace implements WebspaceInterface
+class WebspaceExport extends Export implements WebspaceExportInterface
 {
-    /**
-     * @var EngineInterface
-     */
-    protected $templating;
-
-    /**
-     * @var DocumentManager
-     */
-    protected $documentManager;
-
     /**
      * @var EngineInterface
      */
@@ -56,9 +45,9 @@ class Webspace implements WebspaceInterface
     protected $extensionManager;
 
     /**
-     * @var ContentExportManagerInterface
+     * @var ExportManagerInterface
      */
-    protected $contentExportManager;
+    protected $exportManager;
 
     /**
      * @var string[]
@@ -66,12 +55,17 @@ class Webspace implements WebspaceInterface
     protected $formatFilePaths;
 
     /**
+     * @var Output
+     */
+    protected $output;
+
+    /**
      * @param EngineInterface $templating
      * @param DocumentManager $documentManager
      * @param DocumentInspector $documentInspector
      * @param StructureManagerInterface $structureManager
      * @param ExtensionManagerInterface $extensionManager
-     * @param ContentExportManagerInterface $contentExportManager
+     * @param ExportManagerInterface $exportManager
      * @param array $formatFilePaths
      */
     public function __construct(
@@ -80,7 +74,7 @@ class Webspace implements WebspaceInterface
         DocumentInspector $documentInspector,
         StructureManagerInterface $structureManager,
         ExtensionManagerInterface $extensionManager,
-        ContentExportManagerInterface $contentExportManager,
+        ExportManagerInterface $exportManager,
         array $formatFilePaths
     ) {
         $this->templating = $templating;
@@ -88,8 +82,9 @@ class Webspace implements WebspaceInterface
         $this->documentInspector = $documentInspector;
         $this->structureManager = $structureManager;
         $this->extensionManager = $extensionManager;
-        $this->contentExportManager = $contentExportManager;
+        $this->exportManager = $exportManager;
         $this->formatFilePaths = $formatFilePaths;
+        $this->output = new NullOutput();
     }
 
     /**
@@ -104,13 +99,21 @@ class Webspace implements WebspaceInterface
         $nodes = null,
         $ignoredNodes = null
     ) {
+        $this->exportLocale = $locale;
+        $this->output = $output;
+        $this->format = $format;
+
+        if (null === $this->output) {
+            $this->output = new NullOutput();
+        }
+
         if (!$webspaceKey || !$locale) {
             throw new \Exception(sprintf('Invalid parameters for export "%s (%s)"', $webspaceKey, $locale));
         }
 
         return $this->templating->render(
-            $this->getTemplate($format),
-            $this->getExportData($webspaceKey, $locale, $output, $format, $uuid, $nodes, $ignoredNodes)
+            $this->getTemplate($this->format),
+            $this->getExportData($webspaceKey, $uuid, $nodes, $ignoredNodes)
         );
     }
 
@@ -119,31 +122,24 @@ class Webspace implements WebspaceInterface
      */
     public function getExportData(
         $webspaceKey,
-        $locale,
-        $output = null,
-        $format = '1.2.xliff',
         $uuid = null,
         $nodes = null,
         $ignoredNodes = null
     ) {
         /** @var PageDocument[] $documents */
-        $documents = $this->getDocuments($webspaceKey, $locale, $uuid, $nodes, $ignoredNodes);
+        $documents = $this->getDocuments($webspaceKey, $uuid, $nodes, $ignoredNodes);
         /** @var PageDocument[] $loadedDocuments */
         $documentData = [];
 
-        if (null === $output) {
-            $output = new NullOutput();
-        }
+        $this->output->writeln('<info>Loading Data…</info>');
 
-        $output->writeln('<info>Loading Data…</info>');
-
-        $progress = new ProgressBar($output, count($documents));
+        $progress = new ProgressBar($this->output, count($documents));
         $progress->start();
 
         foreach ($documents as $key => $document) {
-            $contentData = $this->getContentData($document, $locale, $format);
-            $extensionData = $this->getExtensionData($document, $format);
-            $settingData = $this->getSettingData($document, $format);
+            $contentData = $this->getContentData($document, $this->exportLocale, $this->format);
+            $extensionData = $this->getExtensionData($document, $this->format);
+            $settingData = $this->getSettingData($document, $this->format);
 
             $documentData[] = [
                 'uuid' => $document->getUuid(),
@@ -158,166 +154,17 @@ class Webspace implements WebspaceInterface
 
         $progress->finish();
 
-        $output->writeln([
+        $this->output->writeln([
             '',
             '<info>Render Xliff…</info>',
         ]);
 
         return [
             'webspaceKey' => $webspaceKey,
-            'locale' => $locale,
-            'format' => $format,
+            'locale' => $this->exportLocale,
+            'format' => $this->format,
             'documents' => $documentData,
         ];
-    }
-
-    /**
-     * Returns a array of the given content data of the document.
-     *
-     * @param BasePageDocument $document
-     * @param $locale
-     * @param $format
-     *
-     * @return array
-     */
-    protected function getContentData(BasePageDocument $document, $locale, $format)
-    {
-        /** @var BasePageDocument $loadedDocument */
-        $loadedDocument = $this->documentManager->find($document->getUuid(), $locale);
-
-        /** @var \Sulu\Component\Content\Metadata\StructureMetadata $metaData */
-        $metaData = $this->documentInspector->getStructureMetadata($document);
-
-        $propertyValues = $loadedDocument->getStructure()->toArray();
-        $properties = $metaData->getProperties();
-
-        $contentData = $this->getPropertiesContentData($properties, $propertyValues, $format);
-
-        return $contentData;
-    }
-
-    /**
-     * Returns the Content as a flat array.
-     *
-     * @param PropertyMetadata[] $properties
-     * @param $propertyValues
-     * @param $format
-     *
-     * @return array
-     */
-    protected function getPropertiesContentData($properties, $propertyValues, $format)
-    {
-        $contentData = [];
-
-        foreach ($properties as $property) {
-            if ($this->contentExportManager->hasExport($property->getType(), $format)) {
-                if (!isset($propertyValues[$property->getName()])) {
-                    continue;
-                }
-
-                $propertyValue = $propertyValues[$property->getName()];
-
-                if ($property instanceof BlockMetadata) {
-                    $data = $this->getBlockPropertyData($property, $propertyValue, $format);
-                } else {
-                    $data = $this->getPropertyData($property, $propertyValue, $format);
-                }
-
-                $contentData[$property->getName()] = $data;
-            }
-        }
-
-        return $contentData;
-    }
-
-    /**
-     * Creates and returns a property-array.
-     *
-     * @param PropertyMetadata $property
-     * @param PropertyValue $propertyValue
-     * @param string $format
-     *
-     * @return array
-     */
-    protected function getPropertyData(PropertyMetadata $property, $propertyValue, $format)
-    {
-        return $this->createProperty(
-            $property->getName(),
-            $this->contentExportManager->export($property->getType(), $propertyValue),
-            $this->contentExportManager->getOptions($property->getType(), $format),
-            $property->getType()
-        );
-    }
-
-    /**
-     * Creates and Returns a property-array for content-type Block.
-     *
-     * @param BlockMetadata $property
-     * @param PropertyValue $propertyValue
-     * @param $format
-     *
-     * @return array
-     */
-    protected function getBlockPropertyData(BlockMetadata $property, $propertyValue, $format)
-    {
-        $children = [];
-
-        $blockDataList = $this->contentExportManager->export($property->getType(), $propertyValue);
-
-        foreach ($blockDataList as $blockData) {
-            $blockType = $blockData['type'];
-
-            $block = $this->getPropertiesContentData(
-                $property->getComponentByName($blockType)->getChildren(),
-                $blockData,
-                $format
-            );
-
-            $block['type'] = $this->createProperty(
-                'type',
-                $blockType,
-                $this->contentExportManager->getOptions($property->getType(), $format),
-                $property->getType() . '_type'
-            );
-
-            $children[] = $block;
-        }
-
-        return $this->createProperty(
-            $property->getName(),
-            null,
-            $this->contentExportManager->getOptions($property->getType(), $format),
-            $property->getType(),
-            $children
-        );
-    }
-
-    /**
-     * Returns a array with the given value (name, value and options).
-     *
-     * @param $name
-     * @param $value
-     * @param array $options
-     * @param string $type
-     * @param array $children
-     *
-     * @return array
-     */
-    protected function createProperty($name, $value = null, $options = [], $type = '', $children = null)
-    {
-        $property = [
-            'name' => $name,
-            'type' => $type,
-            'options' => $options,
-        ];
-
-        if ($children) {
-            $property['children'] = $children;
-        } else {
-            $property['value'] = $value;
-        }
-
-        return $property;
     }
 
     /**
@@ -328,7 +175,7 @@ class Webspace implements WebspaceInterface
      *
      * @return array
      */
-    protected function getExtensionData(BasePageDocument $document, $format)
+    protected function getExtensionData(BasePageDocument $document)
     {
         $extensionData = [];
 
@@ -337,7 +184,7 @@ class Webspace implements WebspaceInterface
             $extension = $this->extensionManager->getExtension($document->getStructureType(), $extensionName);
 
             if ($extension instanceof ExportExtensionInterface) {
-                $extensionData[$extensionName] = $extension->export($extensionProperties, $format);
+                $extensionData[$extensionName] = $extension->export($extensionProperties, $this->format);
             }
         }
 
@@ -352,7 +199,7 @@ class Webspace implements WebspaceInterface
      *
      * @return array
      */
-    protected function getSettingData(BasePageDocument $document, $format)
+    protected function getSettingData(BasePageDocument $document)
     {
         if ($created = $document->getCreated()) {
             $created = $created->format('c');
@@ -367,7 +214,7 @@ class Webspace implements WebspaceInterface
         }
 
         $settingOptions = [];
-        if ($format === '1.2.xliff') {
+        if ($this->format === '1.2.xliff') {
             $settingOptions = ['translate' => false];
         }
 
@@ -430,14 +277,13 @@ class Webspace implements WebspaceInterface
      */
     protected function getDocuments(
         $webspaceKey,
-        $locale,
         $uuid = null,
         $nodes = null,
         $ignoredNodes = null
     ) {
-        $queryString = $this->getDocumentsQueryString($webspaceKey, $locale, $uuid, $nodes, $ignoredNodes);
+        $queryString = $this->getDocumentsQueryString($webspaceKey, $uuid, $nodes, $ignoredNodes);
 
-        $query = $this->documentManager->createQuery($queryString);
+        $query = $this->documentManager->createQuery($queryString, $this->exportLocale);
 
         return $query->execute();
     }
@@ -455,7 +301,6 @@ class Webspace implements WebspaceInterface
      */
     protected function getDocumentsQueryString(
         $webspaceKey,
-        $locale,
         $uuid = null,
         $nodes = null,
         $ignoredNodes = null
@@ -475,7 +320,7 @@ class Webspace implements WebspaceInterface
         // filter by locale
         $where[] = sprintf(
             '[i18n:%s-template] IS NOT NULL',
-            $locale
+            $this->exportLocale
         );
 
         // filter by uuid
@@ -484,6 +329,7 @@ class Webspace implements WebspaceInterface
         }
 
         $nodeWhere = $this->buildNodeUuidToPathWhere($nodes, false);
+
         if ($nodeWhere) {
             $where[] = $nodeWhere;
         }
@@ -552,29 +398,5 @@ class Webspace implements WebspaceInterface
         }
 
         return $paths;
-    }
-
-    /**
-     * Returns export template for given format like XLIFF1.2.
-     *
-     * @param $format
-     *
-     * @return string
-     *
-     * @throws \Exception
-     */
-    protected function getTemplate($format)
-    {
-        if (!isset($this->formatFilePaths[$format])) {
-            throw new \Exception(sprintf('No format "%s" configured for webspace export', $format));
-        }
-
-        $templatePath = $this->formatFilePaths[$format];
-
-        if (!$this->templating->exists($templatePath)) {
-            throw new \Exception(sprintf('No template file "%s" found for webspace export', $format));
-        }
-
-        return $templatePath;
     }
 }