+++ src/Sulu/Component/Import/Manager/ImportManager.php
@@ -9,7 +9,7 @@
  * with this source code in the file LICENSE.
  */
 
-namespace Sulu\Component\Content\Import;
+namespace Sulu\Component\Import\Manager;
 
 use PHPCR\NodeInterface;
 use Sulu\Component\Content\Compat\PropertyInterface;
@@ -18,9 +18,9 @@ use Sulu\Component\Content\ContentTypeManagerInterface;
 use Sulu\Component\Content\Import\Exception\ContentTypeImportMissingException;
 
 /**
- * Import content by given xliff file from ContentExportManager.
+ * Import content by given xliff file.
  */
-class ContentImportManager implements ContentImportManagerInterface
+class ImportManager implements ImportManagerInterface
 {
     /**
      * @var ContentTypeManagerInterface